PYIC-5675: Add SessionUserIssuedCredentialsTable
This adds a new DynamoDB table for storing users credentials gathered
during their session, before writing them to a long term store. This
change is just to create the table, and not start writing to it.

It also adds a class to represent the data that will be stored in it -
I'm adding this with the table to ensure the correct format of the data,
specifically the sort key. The class may change slightly when we
acutally implement the writes.