PYIC-5685: Store credentials in SessionCredentialsTable
This adds logic to start storing VCs received in the new
SessionCredentialsTable. It creates a new service to do this, and wires
it into the CriStoringService.

It also updates our VerifiableCredential class to store the parsed
credential alongside the raw format. This will allow us to easily access
the signature when needed. We've already parsed it so might as well keep
it.