PYIC-5685: Store VC in session credential table on reuse
If a user has already proved their identity, we want to move their VC to
the session credentials table. This is so that when we come to build
their identity, the VC required to do so are in the same place as if
they'd just completed a full identity proving journey.