PYIC-5685: Remove creds from session credential store in reset-identity
This updates the reset-identity handler to remove any VC from the
session credential table for the users current session, as well as from
the tactical store.

I'm a bit dubious that there will ever actually be anything to remove
here. If we're resetting the users identity they'll have failed to match
a profile and so won't have anything in the session credentials store
yet. And this happens at the start of their journey, so they shouldn't
have visited any CRI yet.