BAU: Move ConfigService from BasicEvent to JourneyContext
This change means that BasicEvents can now be deserialized from the
journey maps without creating any ConfigService's. This was making it
hard to test, as we couldn't intercept and mock those ConfigServices.
They were trying to establish connections with AWS which was failing in
the pipelines.

By moving the ConfigService to the journey context, we can deserialize,
and mock, without issue.