PYIC-5685: Refactor reuse credential persistance
This moves the logic that stores the credentials in a reuse journey
closer to where we determine which type of reuse journey the user is on.

This allows us to be a bit more accurate about which credentials we
store and the `receivedThisSessionFlag`.

For an operational profile reuse scenario, we can correctly set the flag
when a user is in migration - they will have received the operational
proile VC in the same session.

For a 6mfc journey we can just store the required credentials.
Previously we were storing all of them and would have had to remove the
fraud VC. Now it's just one operation.

For other reuse journeys we can store the credentials related to the
profile met - in practice this will just be gpg45 credentials.