PYIC-5685: Allow DataStore batch deletes
The `delete(String partition key)` method on the data store implied that
it would delete all items that had that partition key. It actually would
have only worked on a table that only defined a partition key as part of
it's schema. Use on any table that also used a hash key would result in
an error.

This updates that method to now remove all items that use that partition
key. It fetches all items, and then adds them all to a batch delete
operation.