PYIC-5733: Read from session cred table in TicfCriService
This allows the TicfCriService to fetch VC from the new session
credentials table, or the user issued credentials table, depending on
the flag.

It adds a new signature on the SessionCredentialsService that allows for
fetching credentials based on if they were received this session or not.
This required a modification to an unused method in the DataStore.