PYIC-5733: Add getLatestCredentialsByIssuer to SessionCredentialsService
There is a scenario where a user may have more than one address VC in
the session-credentials-table for there current session. If they are on
a 6MFC jouurney we will have copied their address VC into the table. If
they then let us know they want to update it they will be sent to the
address CRI and collect another address VC.

This is only really a problem when building shared attributes for other
CRIs - we only want to include the newest address.

This adds a method to the sessionCredentialsService class which will
return the newest (by not before date) VC for each CRI for the users
current session from the session credentials table. We then use that in
the build-cri-oauth-request lambda.