PYIC-5723: Read from session cred table in BuildProvenUserIdentity
This updates the build-proven-user-identity lambda to read from the
session credentials table if the flag is set.

If we are reading from the new table, we no longer filter the VC by
profile type. This shouldn't be neccessary as we no longer display the
reuse page for operational profile reuse journeys - the user is just
sent straight back to the service. And when we _do_ show this page (for
a gpg45 reuse journey), any operational profile VC should have already
been removed for the user if previously present.

This also gets all the credentials from the table, rather than the
"latest by CRI", as we should have only stored and/or fetched the latest
versions in the long term table, rather than possibly having a duplicate
address VC.