PYIC-5733: Read from session cred table in BuildUserIdentityHandler
This allows the build-user-identity-lambda to read from the new session
credentials table, depending on the status of the flag.

It uses the `getLatestCredentialsByIssuer()` method, as it's important
that we only use the newest address VC in the case of a 6MFC journey.

This also refactors the ErroResponse class, and the
HttpResponseExceptionWithErrorBody to make it easier to return an error
response.