PYIC-5851: Implement store-identity lambda
This adds the logic to the new store-identity lambda.

It fetches the credentials for the users current session from the
session credentials table (the most recent by CRI), and then stores
them.

To store them there is a new method on the VerifiableCredentialService
that first clears all the users VC, and then writes the current session
ones.