PYIC-5851: Wire in the store-identity lambda
This adds some new states for the store-identity lambda, and wires them
in to the journey map before the success page, and before the f2f
handoff page.

There's an interesting wrinkle due to having to take both the
sessionCredentialsTableReads and ticfCriBeta feature flags. The EVALUATE_GPG45_SCORES
state uses both. The process-journey-event lambda will use the target
state from the first enabled feature flag it finds. I believe the way
I've set this up then we should be able to use both flags independently
or combined and get the desired effect.