PYIC-5733: Remove repeat address VC from session creds table
When a user runs a 6MFC journey we copy all their VC expect the expired
fraud VC into the session credentials table, including their address
VC.

If, on the journey, the user tells us they need to update their address
we send them to the address CRI where they collect another address VC.

To avoid confusion, specifically in the build-cri-oauth-request and
build-user-identity lambdas, we have decided to remove the old address
VC before storing the new one. This keeps the session creds table VC's
the users "working" identity.

We may in future decide to extend this logic to other CRIs. However
currently we only have the need for the address CRI so we're keeping it
scoped.

This change means we no longer need to be able to fetch the "latest" VC
for each CRI so can remove some logic that was handling this.