BAU: Recreate event mapping
Previously we were enabling and disabling the event mapping that linked
an SQS queue to the process async lambda. There seems to be an issue
with this. Often, after renabling, the trigger no longer works.
This takes a different approach and removes the event mapping
completelely, and then recreates it.