PYIC-5635: Upgrade AWS SDK from v1 to v2
There are quite a few changes here, but the biggest part is replacing
the sevice clients with new version, and building the model objects
(mostly request and responses for/from the services) with builders.

For the clients, we now specify the region and the credentials providers
explicitly. This is recommended, as it prevents AWS having to look these
things up and slowing things down.

This also uses the BOM for the sdk, rather than a shared version for
each service sdk. The AWS docs state that whilst the versions across all
services do currently match, that can't be guaranteed in future. Using
the BOM should protect us if that changes.