PYIC-5635: Don't set credentialsProvider on AWS clients
The docs for upgrading from v1 to v2 of the SDK suggest that if you're
using lambdas you should explicitly set the credentialsProvider to env
vars, which are provided by the lambda runtime. This is to avoid the
credential provider chain having to search for credentials in different
places, and makes things faster.

This doesn't work if you're using snapstart. Using snapstarts causes the
runtime to use container credentials instead, so no env vars can be
found. Doing this prevents the credentials from expiring before the
function is restored.

https://docs.aws.amazon.com/lambda/latest/dg/snapstart-activate.html#snapstart-credentials