PYIC-5635: Use UrlConnectionHttpClient for sdk clients
The docs suggest that for lambdas you're better to use the more
lightweight UrlConnectionHttpClient, than the default. There is an async
version too, but I'm not sure we need that.

This updates the clients to use the lightweight sync http client. The
recommendation is to pass the sdk client builder a builder for the http
client. This allows the sdk to manage the http clients lifecycle.

This also adds some config to the root gradle project to exclude the
other http clients brought in by the sdk, to keep the classpath smaller
and lighter.

https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/lambda-optimize-starttime.html