PYIC-5948: Filter VC before checking gpg45 match
We had a bug where if a user completed a F2F journey, but failed without
a CI (no validity score), and then followed that with a hmrc migration
journey with evidence, they could incorrectly match an M1A profile.

This is because we weren't filtering out the operational profile when
doing gpg45 evaluation. The validity score from the migration VC was
contributing and allowing the user to match M1A.