PYIC-5932: Handle not found mitigation routes
We have logic that finds a mitigation route for a users CI based on the
dcoument type associated with the CI. Previously we were throwing an
error if a mitigation route couldn't be found. This was because we
assumed that there were no circumstances where a particular CI could be
issued that we wouldn't be able to mitigate.

The NINO CRI has broken that rule - we don't mitigate a particular CI
from NINO. So rather than throwing, we should just not find a mitigation
route and allow the usual failure event to be issued instead.