PYIC-6028: Stop dual writing credentials
This stops core from dual writing VCs to the session-credentials table
and the user-issued-credentials table.

This was previously covered in [PR#1845](https://github.com/govuk-one-login/ipv-core-back/pull/1845),
however I've noticed that deploying that PR with live traffic results in
errors.

It appears the permissions removal on that PR happens before the new
version of the process-cri-callback lambda is active - with about a 4
minute window where the outgoing lambda is trying to write to the
user-issued-credentials table but without the correct permissions,
resulting in a technical error. This could affect a lot of users.

I may split that PR further to ensure all permission changes go out
separately to the code changes that require them.