PYIC-5966: Send IPV_IDENTITY_STORED audit event
This adds a new audit event for when we store an identity from the
session credentials table into the long term storage solution.

It includes an extension for the vot. We won't always have one however,
specifically in the f2f journey. We store the partial identity before
the f2f handoff page, so that the VCs are in the long term store ready
for when the async VC arrives. In that scenario, we're setting the vot
to null.