PYIC-6200: Add coi-check lambda
This adds the coi-check lambda, which will ensure that for COI journeys
the old identity matches the new identity. This is done by checking the
names and DOB from both the new and old.

This change updates and reuses the correlation checking logic already
available in the user identity service.

There will need to be subsequent PRs to wire the new lambda in.