PYIC-6397: Add back event to update details page
The update details page now has a back link. That link should take the
user back to the reuse landing page.

There are other ways to get to the update details page, but they are all
"back" events from pages a user would reach directly after the update
details page, so we don't want the back link to ever return the user to
those pages.