BAU: Fix back after repeat fraud check
When a user was on a 6MFC journey they could choose to go back to check
their details again, after having said they wanted to change their
details.

That was routing them to the COI journey start, which is incorrect.

This introduces a couple of extra states so we can handle the back
event correctly.