PYIC-6088: Refactor to handle scopes
This introduces the use of the nimbus Scope object. Scopes can be a
space separated list of values. And that is in fact what we already
receive from the orchestrator. Using the object will allow us to to more
advanced things in the future.

This also adds a case when validating the scope in the JAR request. We
should never recieve both openid and reverification in the same request.