PYIC-6088: Move scope validation into claimsSet validation
This pushes the specific validation of the scope claim further into the
JarValidator. This allows us to throw an exception and catch it to be
rethrown as a recoverable exception in line with how we do it elsewhere.

This also adds scope as a required claim to the Jar validator.