PYIC-6088: Use Instant when validating JAR ttl
Instant.now() will return a time in UTC. We were using LocalDateTime and
OffsetDateTime which are timezone unaware, so will differ depending on
the system clock that is generating them.

This was manifesting with tests passing locally but not in GHA. Local
tests were getting a BST time, and GHA were getting a UTC time.