PYIC-6088: Fix scope validation
We were validating that all scopes provided were defined in config for a
given client. This was too strict. Orch gives is more than just
`openid`, which is the oonly scope we care about for them.

This updates the logic to ensure that only clients configured for openid
can use it, and only clients configured for reverification can use it.
If the client doesn't give us either of these we log it and allow an IPV
journey. This is due to the values in the scope being passed on to us
from them by orch, and we don't want to block unneccessarily. If we see
logs showing neither provided we can address this.