PYIC-6424: Set VOT to P0 in reset-session-identity lambda
This sets the users VOT to P0 at the point we reset some or all of the
VCs in the session credential store.

This is to reflect the fact that the VCs in the store no longer form a
valid P2 identity.

This is especially important for COI journeys where a user has an
existing valid identity they're reusing. Their VOT will have already
been set to P2. If they choose to update their details, but the fall out
of COI journey either by abandoning a CRI or a fail-with-no-ci, we need
to make sure that the VCs they are returned to the service with match
the VOT in their session.

There is probably more work to do around content, or potentially sending
the user back to the check-existing-identity lambda to have another go,
but this is an MVP solution.