PYIC-6579: Remove COI check from address only journeys
If a user is updating just their address then there is no need to run a
COI check. Their name and DOB won't have changed.

This also removes the actual check from the lambda and replaces it with
some logging. We should never see that log.

This leaves the ADDRESS_ONLY enum in place - it's used by the reset
session lambda to remove the correct VCs. This will probably get
addressed in an upcoming ticket to use contexts instead of a value in
session.