PYIC-6582: Use context for coi-check and reset-session-identity lambdas
This moves from using a `coiSubjourneyType` value in the users session
to injecting the required reset or check type in the journey map.

This means we're not relying on state that gets set by magic in a
different lambda, which could be error prone.

This removes the existing enum as it wasn't really appropriate for both
contexts, and replaces it with more specific versions. The
reset-session-identity lambda is used in more than just COI journeys for
example.