Revert "PYIC-5872: Open API gateway"
This reverts all the work done to open core's internal API gateway.
There is a problem with the SAM template when deploying to non dev
accounts.

It appears that CloudFormation tries to resolve the reference to the
conditionally created new API gateway in the function functino events,
before it checks if those events are even being created. It fails to
create the CloudFormation change set with the following:

Template format error: Unresolved resource dependencies [IPVCoreInternalTestingApi] in the Resources block of the template