PYIC-5872: Use lambda perm over conditional event
Having conditional events on functions wasn't working in non dev
environments. It appears CloudFormation wanted to resolve the nested
reference to the conditionally created API gateway before noticing that
the event itself didn't need creating. This led to errors due to an
unresolveable reference.

Having inspected the CloudFormation stacks created with and without the
event, the only extra resource being created was the permission on the
lambda to allow the testing API gateway to invoke it. All the other API
gateway resources that it might create were already handled by using the
OpenApi spec when creating the gateway itself.

We can create that permission separately and use a conditional for it.