PYIC-6581: Start storing users states in a stack
This change starts recording the users transition through the journey
states in a stack. This is to allow us to support a native "back"
functionality, rather than having to have specific events for it.

This includes a new convertor for the stack - dynamo can't save a
complex type without you telling it how to do it.

We need to merge this before any changes that might start consuming the
stack to ensure that all active sessions have a stack to work with.