PYIC-6634: Add IPV_USER_DETAILS_UPDATE_ABORTED audit event
Fired when a user choosed not to continue with a COI journey. This can
happen by continuing back to the service without changing anything, or
going back to a previous state.

For the 6MFC journey going back from the name change state will trigger
this event, where as the reuse journeys don't. This is because for 6MFC
journeys going back takes the user to the confirm-your-details which
isn't actually part of the COI journeys. For the reuse COI journeys
going back takes them to a COI specific page - update-details. The back
action from here will fire this event, so it looks a bit tidier in the
journey maps.