PYIC-6724: Start reading user state from stack
This switches core-back to start using the users state and subjourney
type from the stack, rather than the individual fields in their session.
It removes those fields entirely.
This also updates the exception thrown if the stack is empty to a
runtime exception. Keeping it as checked would have made the
process-journey-event lambda excessivley verbose. And this is an
exception that shouldn't ever happen in reality.