PYIC-6796: Add unsupported COI events to journey map
Fraud are interested in the different combinations of attributes that
are unsupported that a user may try to change. This change handles new
events from core-front that give us that info for the audit events.

This leaves in place the names-dob event which will become deprectated
once core-front starts sending these new events, and will be removed
after that. Even though these are all behind a feature flag, it'd be
nice to not break things.