PYIC-6724: Implement back functionality in journey-engine
This adds functionality to the journey engine to detect a back event,
and find a legitimate previous state for them from the stack.

Initially we will only support back where the current state and the
previous state are page states.

This change needs to take account of jumps between subjourneys, as well
as nested journeys.

This change removes the addition of JourneyChangeStates to the stack -
they make the logic a lot more complicated and add little of value.

This change also adds a new field to the states object - journeyType.
This in populated when the state machine is initialized, and makes
writing states to the stack a lot simpler, particularly when moving
between subjourneys.

When a user does go "back", this change actually adds the previous state
on top of the stack, rather than just popping off the top one. This will
give us a fuller picture of the users journey should we need it.