PYIC-6724: Default to back event on state
If a state has a back event explicitly defined we should use it rather
than the inbuilt mechanism. The motivation for this is that we currently
have back events defined that emit audit events defined in the journey
map. If we just use the inbuilt functionality we'd lose the audit
events.

This will also allow us to configure a back event that does something
other than go back to the user previous state, or to a non page state.
Such as in the passport mitigation journey.