PYIC-6870: Remove coiEnabled flag and redundant contexts
This removes the use of the coiEnabled feature flag in the journey maps.
We've been able to drop a couple of states because of this.

There are also some now unneccessary contexts which can also be removed.

This also adds a new event to the `IDENTITY_REUSE_PAGE_TEST` state. This
is to get the tests to pass. We test that page states all have the same
events defined on them. The removal of the flag tripped the test up.
This state is a test state for a feature we aren't using, so I've just
routed to the technical error page.

There will be a fronend-change which should be merged before this.

f