BAU: Fix orphaned states in the journey map
The orphaned states option wasn't working. When it was created it
specifically excluded the initial journey state from it's output, as it
legitimately had no events pointing at it. This was one we had just one
map though.

Now we've split the map, we have lots of "start" states. Keeping track
of and excluding them would be hard. This option is still useful, as
it's easy to see which states are legitimate once applied, highlighting
any erroneous states.