PYIC-5875: Define value for internal testing API key
The previous approach to creating and then extracting the API key for
the internal testing API wan't working. The role being created to allow
GHA to fetch the API key had to have a permissions bounary associated
with it. That boundary doesn't allow the describe-stack or get-api-key
actions.

Rather than trying to get those permissions added, which might be
difficult and long, we can instead provide a value for the key manually,
stored in secrets manager.

GHA will be able to fetch that key with an API call. And if it can't, we
can just set the API key as a GHA secret.