PYIC-6854: Namespace dev env test API key
You can't have two API keys in API Gateway that have the same value.
This means the approaching of using the same secret for all dev env API
keys doesn't work as they'll clash.

This namespaces the API key in the dev accounts with the environment
which should prevent a clash.