PYIC-6961: Return build-client-oauth-response before checking session
If a user has lost their session when returning from a CRI (very
possibly because the redirect has opened a different browser), then we
show them a recoverable error page with a continue button. That button
fires a `build-client-oauth-response` event into core-back.

Core-back should see that event and return a
`build-client-oauth-response` journey, which the step function uses to
invoke the lambda to actually close the OAuth session with the client.

Instead we've been throwing an error due to the ipvSessionId not
existing.

This changes the logic to return that special case response before we
try to check the session ID.