PYIC-6883: Use async client for Audit service
This adds the option to use an async SQS client for the Audit Service.

Inititally we want this behind a feature flag, in case we find any
issues with it. This requires having two clients ready. Hopefully this
won't add too much overhead as they'll only be initialized when used for
the first time.