PYIC-6083: Store inherited identities in EVCS
This updates the initalise session lambda to store inherited identities
in EVCS, rather than the tactical store.

We're not going to be going live with inherited identities before we're
migrated over to EVCS, so we don't need to worry about migrating
anything.

Adds a new method to the EVCS service to do the updating of an existing
inherited ID to historic, and storing a new one as current.