PYIC-6083: Don't update inherited identity when storing P2
We want a user to be able to have an inherited identity along side a P2
identity. This means that if a user with an existing inherited identity
attains a P2, we don't want to lose the inherited identity VC.

This adds metadata to the inherited identity VC in EVCS, which we can
then use when storing completed identities to ensure we don't update the
inherited identity from `CURRENT` to `HISTORIC`.

There are other options for how to do this. We could parse the VCs and
check the issuer, but this is a bit messy and doesn't play nicely with
the streams we're using.

The metadata itself could be in a different shape. This uses the key
"inheritedIdentity" with a value of the Cri ID. This may be more useful
in future that just a boolean flag.