PYIC-7022: Include current inherited identity VC in bundle
When we store an inherited identity VC in EVCS it has a status of
`CURRENT`. If a user is returning from a F2F journey we were only using
the VCs with a `PENDING_RETURN` status. We need to ensure we're not
throwing away the current inherited identity VC, so that we can match
the correct VOT in the VTR for inherited identity journeys.