PYIC-7047: Standardise secret names
This updates the names of our secrets in secrets manager to be in line
with how we name our other parameters.

This has allowed us to drop the specific getApiKeySecret from the config
service, as all the api keys are now stored as plaintext rather than in
a json blob.

The new secrets have all been added in the higher envs, and there is a
separate PR on the config repo to add them in the lower envs. This
should not go in before that.