PYIC-6083: Handle multiple inherited identities
It should never be the case that a user has more than one current
inherited identity in the EVCS. We can handle that scenario gracefully
however if we do encounter it. This will move the existing VCs to
historic if the incoming is stronger or equal to all of them. If it's
weaker we do nothing, and the multiple existing will be used when
matching the requested VOT.