PYIC-7324: Write VCs directly to EVCS for reuse tests
API tests that required an existing identity would have to run a full
journey through the system to generate it. This change allows for VCs
generated by the CRI stubs to be written directly to EVCS for a user.

This should mean less boilerplate as well as a faster running test
suite.