PYIC-7323: Add app test journeys
This adds a happy path journey for a BRC check (subject and evidence
details taken from existing selenium tests).

It also adds a test for dcmaw returning a 404 when attempting to fetch
the users VC. I've added a new step rather than extending the existing
error request step, for clarity.