PYIC-7470: Add bulk migration batchIds to tactical store VCs
This adds the batch ID from the bulk migration to the VCs in the
tactical store. This allows for us to more easily deal with having to
roll back a batch.

The Bravo team will be responsible for deleting VCs from EVCS. We can
then run the script again and remigrate the VCs whos batch IDs match a
value in config.

And for on-demand migrations in the check-existing-identity lambda we
can follow similar logic - check if the users tactical VCs have a
matching batch ID, and remigrate if they do.