PYIC-7470: Recreate EvcsClient in bulk migration
Testing of the bulk migration script has shown that we're getting
predicatable GOAWAY responses from the API gateway in front of the EVCS
stub.

These are happening in clusters at a cadence of every 11500 - 11750
identities migrated. I can't find anything in AWS docs to directly
suggest the reason, however I did note that an ALB will do this if it
receives over 10'000 requests from any given connection. So I suspect
there is a similar mechanism happening here.

To resolve this, we can generate a new EVCS client for each page of
results we're processing. This should keep us well under the figure
noted above. There doesn't seem to be any way to force the client to
drop any open connections easily, so recreating might be the easiest.

I did consider adding a GOAWAY response as a valid reason for a retry,
in the retry logic in in the client, but didn't want to impact the core
system code.