PYIC-7470: Only recreate EVCS client when needed
Recreating the EVCS client solved the GOAWAY issue, but it introduced a
new one. The audit service was unable to open a socket due to too many
files being open.

This updates the logic to only recreate the client when the number of
migrated identities is approaching 10_000. This should dramatically
reduce the number of client created.

There doesn't seem to be an easy way to close an HttpClient in Java 17
(in 21 it's been made AutoCloseable), which would have been nicer.

This also handles errors thrown by the AuditService to ensure that we
get the report back from the lambda. The errors will be logged in the
application logs.