BAU: Avoid logging subject identifier in validation failure
When we receive a VC we check that the subject in it exactly matches the
subject we received from orchestrator. If they don't match we throw an
error. The nimbus lib will log the difference between any claims that
we've configured to be exact matches.

We because aware of this after a CRI team asked us to investigate a
failure in staging. The failure was due to mismatching subjects and it
triggered the identifiers to be logged.

I've searched in Splunk for if this has ever occured in prod - and it
hasn't. This change should prevent us from logging any PII should we
ever receive an unexpected subject.