PYIC-7476: Allow nested journey exit events behind conditionals
The state manchine initializer wasn't correctly initializing exit nested
journey events that were defined behind journey map conditionals
(checkIfDisabled etc). This change makes sure that they are.

It also fixes a couple of places in the journey map where we were trying
to exit to states defined outside of the nested journey definition. It
now uses an exit event to route to the expected state.