PYIC-7476: Add cimit-stub client
We have the ability to send VCs directly to the EVCS stub to allow quick
creation of existing identities. This however skips sending the VCs to
CIMIT.

This adds a new client for doing that. It doesn't use the management API
of CIMIT and instead posts directly into the internal API.

This allows for quick creation of separate session mitigation journeys.