PYIC-7476: Route users to correct page when F2F disabled
If a user backs out of the passport or driving licence CRIs they return
to us with an access-denied event. We show them a page with a couple of
options - either use the passport/driving licence CRI (whichever they
haven't already tried), or try "another way". Which leads to F2F.

If F2F was disabled we were sending users back to the multidoc page
rather than showing them the 'another type of photo ID' page which would
be a bit confusing. Better to let them see this page, and then handle
F2F being disabled if they choose the "another way" option.