PYIC-7507: Add check_details context to auth source check states
The DL CRI need to know if the user is coming to them for a journey that
requires them to check a document, or if it's a journey that's to just
check already gathered document details. This will be established
through the presence of a check_details context claim in the top level
of the jar.