BAU: Update pipeline overview dashoard json
I've made a few more tweaks to the dashboard, so commiting here to keep
in sync.

The biggest change in the query that supplies the honeycombs. There
seems to be an issue where the honeycomb visualisation can't handle a
value of 0 (what we get for a failed deploy). It just decides it doesn't
have enough data, or just shows the previous data point.
This change explicitly maps the binary value to a string, which is then
used as the value to inform the tile. This seems to work and is
potentially more obvious when looking at the query.