PYIC-7533: Expire pending F2F record on failure
When a user fails to prove their identity using F2F for non-CI related
reasons, such as a low score, we show them a error page on their return
journey. This page offers them to try again, or to go back to the
service.

Once they've seen this, we shouldn't need to show it to them again. This
change removes their pending F2F record, so they're not flagged as a
failed F2F jourey on return.

This also solves a bug with an error being thrown if a user has a
mitigated alterate doc CI, and a failed F2F journey.